Check that the docs (output) dir exists

+++ Makefile
@@ -24,7 +24,7 @@ check-links: render
 
 clean:
 	test -n "${OUTPUT_DIR}"
-	find ${OUTPUT_DIR} -depth -name .svn -prune -o -type f -delete
+	if [[ -d ${OUTPUT_DIR} ]]; then find ${OUTPUT_DIR} -depth -name .svn -prune -o -type f -delete; fi
 
 publish-devel: TAG := "head"
 publish-devel: OUTPUT_DIR := $(shell mktemp -d)
